update lista 1

+++ lista 1/Questao2.java
@@ -6,12 +6,12 @@ public class Questao2  {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int counter=0;
-        double soma=0;
+        int counter= 0;
+        double soma= 0;
 
         System.out.println("Informe três notas, em sequência: ");
         while(counter <=2){
-            double nota = sc.nextInt();
+            double nota = sc.nextDouble();
             soma +=nota;
             counter++;
